backend/requirements.txt
amqp==5.0.2
asgiref==3.3.1
billiard==3.6.3.0
celery==5.0.2
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
cryptography==3.2.1
defusedxml==0.7.0rc1
dj-rest-auth==2.0.1
Django==3.1.3
django-allauth==0.43.0
django-cors-headers==3.5.0
django-js-asset==1.2.2
django-mptt==0.11.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
idna==2.10
importlib-metadata==3.1.0
kombu==5.0.2
oauthlib==3.1.0
praw==7.1.0
prawcore==1.5.0
prompt-toolkit==3.0.8
psycopg2==2.8.5
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.4
redis==3.5.3
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-core==3.3.3
sqlparse==0.4.1
stripe==2.55.0
update-checker==0.18.0
urllib3==1.26.1
vine==5.0.0
wcwidth==0.2.5
websocket-client==0.57.0
zipp==3.4.0
